almost done

+++ src/models/task.model.js
@@ -4,15 +4,21 @@ const prisma = new PrismaClient();
 class Task {
   static async createTask(task) {
     try {
-      const newTask = await prisma.task.create({ data: task });
-      return newTask.id;
+      return await prisma.task.create({ data: task });
     } catch (error) {
       throw new Error('Error creating task: ' + error.message);
     }
   }
 
-  static async getUserTasks(userId, page, limit, sortField, sortOrder) {
+  static async getUserTasks(userId, page, sortField, limit, sortOrder) {
     try {
+      const totalTasks = await prisma.task.count({ where: { userId } });
+
+      const validSortFields = ['id', 'title', 'description', 'createdAt', 'updatedAt'];
+      if (!validSortFields.includes(sortField)) {
+        throw new Error(`Invalid ${sortField} field`);
+      }
+
       const tasks = await prisma.task.findMany({ 
         where: { userId },
         orderBy: { [sortField]: sortOrder },
@@ -20,16 +26,15 @@ class Task {
         take: limit 
       });
 
-      return tasks;
+      return { tasks, totalTasks };
     } catch (error) {
       throw new Error('Error fetching user tasks: ' + error.message);
     }
   }
 
-  static async getTaskById(taskId) {
+  static async getTaskById(userId, taskId) {
     try {
-      const task = await prisma.task.findUnique({ where: { id: taskId } });
-      return task;
+      return await prisma.task.findUnique({ where: { id: taskId, userId } });
     } catch (error) {
       throw new Error('Error fetching task by id: ' + error.message);
     }
@@ -37,16 +42,15 @@ class Task {
 
   static async updateTask(taskId, newData) {
     try {
-      await prisma.task.update({ where: { id: taskId }, data: newData });
-      return true;
+      return await prisma.task.update({ where: { id: taskId }, data: newData });
     } catch (error) {
       throw new Error('Error updating task: ' + error.message);
     }
   }
 
-  static async deleteTask(taskId) {
+  static async deleteTask(userId, taskId) {
     try {
-      await prisma.task.delete({ where: { id: taskId } });
+      await prisma.task.delete({ where: { id: taskId, userId } });
       return true;
     } catch (error) {
       throw new Error('Error deleting task: ' + error.message);
